feat: :label: add id field to db types & refactor into firebae and ipfs sub folders
