+++ client/src/app/_components/groups/group-post/GroupPostItem.tsx
@@ -9,8 +9,8 @@ import useStore from '~/stores/utils/useStore';
 import { useUserStore } from '~/providers/store-providers/userStoreProvider';
 import { type UserState } from '~/stores/userStore';
 
-import { type FirebasePostModel } from '~/models/firebase-post-model';
-import { type IPFSPostModel } from '~/models/ipfs-post-model';
+import { type FirebasePostModel } from '~/models/firebase/firebase-post-model';
+import { type IPFSPostModel } from '~/models/ipfs/ipfs-post-model';
 import { api } from '~/trpc/react';
 import BasicButton from '../../ui/BasicButton';
 import { FaHeart, FaRegCommentDots } from 'react-icons/fa6';
