+++ client/src/app/_components/groups/group-post/GroupPostsList.tsx
@@ -4,7 +4,7 @@
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
 import React, { useEffect, useState } from 'react';
 import { api } from '~/trpc/react';
-import { type FirebasePostModel } from '../../../../models/firebase-post-model';
+import { type FirebasePostModel } from '../../../../models/firebase/firebase-post-model';
 import GroupPostItem from './GroupPostItem';
 
 type GroupPostsListProps = {
