+++ client/src/models/firebase/firebase-comment-model.tsx
@@ -1,4 +1,5 @@
 export interface FirebaseCommentModel {
+	id: string;
 	hash: string;
 	posterKey: string;
 }
