+++ client/src/models/firebase/firebase-group-model.tsx
@@ -1,4 +1,5 @@
 export interface FirebaseGroupModel {
+	id: string;
 	name: string;
 	creatorKey: string;
 	groupHash: string;
