+++ client/src/models/firebase/firebase-like-model.tsx
@@ -1,4 +1,5 @@
 export interface FirebaseLikeModel {
+	id: string;
 	hash: string;
 	posterKey: string;
 }
