+++ client/src/models/firebase/firebase-post-model.tsx
@@ -1,4 +1,5 @@
 export interface FirebasePostModel {
+	id: string;
 	hash: string;
 	group: string;
 	posterKey: string;
