+++ client/src/models/firebase/firebase-product-model.tsx
@@ -1,4 +1,5 @@
 export interface FirebaseProductModel {
+	id: string;
 	name: string;
 	creatorKey: string;
 	productHash: string;
