+++ client/src/models/ipfs/ipfs-post-model.tsx
@@ -1,4 +1,5 @@
 export interface IPFSPostModel {
+	id: string;
 	title: string;
 	content: string;
 }
