+++ client/src/models/ipfs/ipfs-product-model.tsx
@@ -1,4 +1,5 @@
 export interface IPFSProductModel {
+	id: string;
 	name: string;
 	price: string;
 	category: string;
